+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
 <!-- Add new, unreleased changes here. -->
+* Added new option: `autoBasePath`. This new flag sets `basePath: true` on all builds. See that option for more details.
 
 ## [3.6.0] - 2017-11-28
 * Added a new lint option: `filesToIgnore`. We'll never report warnings for any
@@ -15,7 +16,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     spelling which will be used preferentially. This was better than making the
     options inconsistent, or spelling the new option `ignoreFiles`.
 * Improved error messages when validation of a polymer.json object fails.
-<!-- Add new, unreleased changes here. -->
 
 ## [3.5.0] - 2017-11-21
 * Added static methods for constructing a ProjectConfig directly from an unvalidated JSON object, in addition to the methods for reading it from disk.